Tile Collision Editor: Allowed using object templates
This change adds a Templates view to the tileset editor, which allows
for selecting templates to place as tile collisions.

Placing tile object templates should still be prevented, because using a
tile from same tileset leads to recursive loading of the same tileset.

TODO: Resolve issues when the same template is selected in both the
      tileset and the map editor and then modified (synchronize, or
      actually use the same dummy document).